Malak/malakproject/requirements.txt
asgiref==3.5.2
boto3==1.24.0
botocore==1.27.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
django-countries==7.3.2
django-phonenumber-field==6.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.8
hvac==0.11.2
idna==3.3
Jinja2==3.1.2
jmespath==1.0.0
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
phonenumbers==8.12.49
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
validators==0.19.0
win32-setctime==1.1.0
